add behavioral tests for logging in.
