+++ .gitignore
@@ -1,6 +1,9 @@
 # Django specific
 staticfiles
 
+# pytest-bdd
+ligapp.tests*
+
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
