+++ dev-requirements.txt
@@ -1,4 +1,5 @@
 pre-commit
+pytest
 pytest-bdd
 pytest-coverage
 pytest-django
