+++ setup.cfg
@@ -3,7 +3,7 @@ line-length=100
 
 [flake8]
 max-line-length=100
-ignore=D102
+ignore=D102,D103
 
 [tool:pytest]
 DJANGO_SETTINGS_MODULE = miniliga.settings
